backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
langchain==0.3.1
langchain-groq==0.2.1
langchain-google-genai==2.0.1
langchain-community==0.3.1
langchain-chroma==0.1.4
langchain-core==0.3.7
google-generativeai==0.8.3
chromadb==0.5.12
tavily-python==0.5.0
pypdf==5.0.1
python-dotenv==1.0.1
httpx==0.27.2
